forgotten int16

+++ nd4j-backends/nd4j-api-parent/nd4j-native-api/src/main/java/org/nd4j/compression/impl/Int16.java
@@ -18,7 +18,7 @@ public class Int16 extends AbstractCompressor {
      */
     @Override
     public String getDescriptor() {
-        return "INT8";
+        return "INT16";
     }
 
     /**
